requirements.txt
python-pptx==0.6.21
python-docx==0.8.11
Pillow==10.2.0
sentence-transformers